+++ copydmdevice/main.c
@@ -6,7 +6,7 @@
 
 int setup_cow_device(const char *, const char *, char *);
 int takedown_cow_device(const char *);
-int extfs_copy(const char *, const char *, int *);
+int extfs_copy(const char *, const char *, unsigned long *);
 
 int main(int argc, char **argv) {
 	const char *dm_device_path;
@@ -15,7 +15,7 @@ int main(int argc, char **argv) {
 
 	char cow_path[PATH_MAX];
 
-	int blocks_copied = 0;
+	unsigned long blocks_copied = 0;
 
 	if (argc != 4) {
 		fprintf(stderr, "usage: %s dm_block_dev dest_path mem_dev\n",
@@ -37,7 +37,7 @@ int main(int argc, char **argv) {
 		/* Don't return on error as we still need to clean up */
 	}
 
-	printf("Copied %d blocks to %s\n", blocks_copied, dest_path);
+	printf("Copied %ld blocks to %s\n", blocks_copied, dest_path);
 
 	if (!takedown_cow_device(dm_device_path)) {
 		error(0, errno, "Error taking down COW device for %s", argv[1]);
